refactor: shared provisioning infrastructure in daemon-common
- DaemonProvisioningConfiguration with @ConditionalOnMissingBean defaults:
  - NoOpEntityScopeProvider — disables MATE interpolation; overridable by
    daemons needing real MATE (Provisiond, Thresholding, Pollerd, etc.)
  - LocalServiceDetectorRegistry — SPI-based, empty unless detector JARs present
- Added mate-api, daemon-loader-shared, detector-registry, provision-api
  deps to daemon-common with ServiceMix exclusions
- Removed duplicated NoOpEntityScopeProvider and bean definitions from
  DiscoveryBootConfiguration
